+++ wolfssl/internal.h
@@ -6737,8 +6737,8 @@ struct WOLFSSL {
     word16          pssAlgo;
 #ifdef WOLFSSL_TLS13
     word16          certHashSigAlgoSz;  /* SigAlgoCert ext length in bytes */
-    byte            certHashSigAlgo[WOLFSSL_MAX_SIGALGO]; /* cert sig/algo to
-                                                           * offer */
+    byte*           certHashSigAlgo;    /* cert sig/algo to offer, allocated
+                                         * when the peer sends the list */
 #endif
 #if defined(HAVE_ECC) || defined(HAVE_ED25519) || defined(HAVE_ED448)
     int             eccVerifyRes;